remove non-necessary casts
git-svn-id: file:///srv/local/data/cleanRepo/trunk@25328 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/main/java/org/matsim/facilities/algorithms/FacilitiesCombine.java
@@ -61,7 +61,7 @@ public class FacilitiesCombine {
 				double cap2 = a2.getCapacity();
 				double cap = a.getCapacity();
 				if ((cap < Integer.MAX_VALUE) && (cap2 < Integer.MAX_VALUE)) { a2.setCapacity(cap + cap2); }
-				else { a2.setCapacity((double) Integer.MAX_VALUE); }
+				else { a2.setCapacity(Integer.MAX_VALUE); }
 			}
 			Iterator<SortedSet<OpeningTime>> ts_it = a.getOpeningTimes().values().iterator();
 			while (ts_it.hasNext()) {
